Debugging

+++ Backend/src/controllers/like.controller.js
@@ -9,7 +9,12 @@ const liked = asyncHandler( async(req, res) => {
     const {blogId} = req.params;
     const userId = req.user?._id;
 
-    if(!blogId || !userId) throw new ApiError(402, "likeController :: like :: UnAuthorized");
+    if(!blogId || !userId) {
+        return res
+                .status(404)
+                .json(new ApiError(404, "Unauthorized Access"));
+    }
+    // if(!blogId || !userId) throw new ApiError(402, "likeController :: like :: UnAuthorized");
 
     const likeDoc = await Like.create(
         {
@@ -20,7 +25,12 @@ const liked = asyncHandler( async(req, res) => {
 
     const likeDocument = await Like.findById(likeDoc._id);
 
-    if(!likeDocument) throw new ApiError(500, "likeController :: like :: Error on registring like");
+    if(!likeDocument) {
+        return res
+                .status(500)
+                .json(new ApiError(500, "Failed to register like"));
+    }
+    // if(!likeDocument) throw new ApiError(500, "likeController :: like :: Error on registring like");
 
     return res
             .status(200)
@@ -34,6 +44,12 @@ const disliked = asyncHandler( async(req, res) => {
 
     const {blogId} = req.params;
     const userId = req.user?._id;
+
+    if(!blogId || !userId) {
+        return res
+                .status(404)
+                .json(new ApiError(404, "Unauthorized Access"));
+    }
 
     await Like.deleteOne(
         {
